change dtype to unin8

+++ bird_detector_paper/ZeroShot.py
@@ -82,7 +82,7 @@ df.to_csv("Figures/annotations.csv",index=False)
 src = rio.open("/orange/ewhite/everglades/Palmyra/palymra.tif")
 numpy_image = src.read()
 numpy_image = np.moveaxis(numpy_image,0,2)
-numpy_image = numpy_image[:,:,:3].astype("float32")
+numpy_image = numpy_image[:,:,:3].astype("uint8")
 
 crop_annotations = deepforest.preprocess.split_raster(numpy_image=numpy_image, annotations_file="Figures/annotations.csv", patch_size=2500, base_dir="crops", image_name="palymra.tif")
 crop_annotations.to_csv("crops/annotations.csv",index=False)
